# Speed up boot times and Improve Json (De)Serialization performance and reduce memory usage by reusing JsonSerializerSettings