--- a/src/Umbraco.Web/PropertyEditors/MultiUrlPickerValueEditor.cs
+++ b/src/Umbraco.Web/PropertyEditors/MultiUrlPickerValueEditor.cs
@@ -125,6 +125,7 @@ namespace Umbraco.Web.PropertyEditors
         {
             NullValueHandling = NullValueHandling.Ignore
         };
+
         public override object FromEditor(ContentPropertyData editorValue, object currentValue)
         {
             var value = editorValue.Value?.ToString();